@TomDonoghue 2021-03-28
Thanks @ryanhammonds - this all looks great!

I checked through the code, and everything looks good - some really nice little clean ups and optimizations here. 
I quick added some small code doc / notes extensions, and extended the test for `fit_fooof_3d` a little because when I was checking I figured it could / should do a little more checking that shapes are conserved as expected. 

Merging this in now!